requirements.txt
fastapi
uvicorn
google-cloud-pubsub
google-cloud-tasks
